[fix] resolve conflict (#CHZZ-44)
resolve conflict

+++ src/main/java/org/chzz/market/domain/bid/repository/BidRepositoryCustomImpl.java
@@ -50,7 +50,7 @@ public class BidRepositoryCustomImpl implements BidRepositoryCustom {
         return jpaQueryFactory
                 .select(new QBiddingRecord(
                         product.name,
-                        auction.minPrice,
+                        auction.minPrice.longValue(),
                         bid.amount,
                         auction.bids.size().longValue(),
                         firstImage.cdnPath,
